+++ test/test_write_read_tags.c
@@ -1562,7 +1562,7 @@ int read_all_tags(TIFF *tif, const TIFFFieldArray *tFieldArray,
                     dblDiffLimit = RATIONAL_EPS * auxDoubleArrayW[i];
                 else
                     dblDiffLimit = 1e-3;
-                dblDiff = auxFloat - auxDoubleArrayW[i];
+                dblDiff = (double)auxFloat - auxDoubleArrayW[i];
                 if (fabs(dblDiff) > fabs(dblDiffLimit))
                 {
                     /*--: EXIFTAG_SUBJECTDISTANCE: LibTiff returns value of
@@ -1573,12 +1573,12 @@ int read_all_tags(TIFF *tif, const TIFFFieldArray *tFieldArray,
                      * indicates special values, which are not treated
                      * within LibTiff!!
                      */
-                    if (!(tTag == EXIFTAG_SUBJECTDISTANCE && auxFloat == -1.0))
+                    if (!(tTag == EXIFTAG_SUBJECTDISTANCE && auxFloat == -1.0F))
                     {
                         fprintf(stderr,
                                 "%u:Read value of %s %f differs from set value "
                                 "%f\n",
-                                i, tFieldName, auxFloat, auxDoubleArrayW[i]);
+                                i, tFieldName, (double)auxFloat, auxDoubleArrayW[i]);
                         GOTOFAILURE
                     }
                 }
@@ -1729,8 +1729,8 @@ int read_all_tags(TIFF *tif, const TIFFFieldArray *tFieldArray,
                                 fprintf(stderr,
                                         "Read value %u of %s #%d %f differs "
                                         "from set value %f\n",
-                                        i, tFieldName, j, auxFloatArray[j],
-                                        auxFloatArrayW[i + (uint32_t)j]);
+                                        i, tFieldName, j, (double)auxFloatArray[j],
+                                        (double)auxFloatArrayW[i + (uint32_t)j]);
                                 GOTOFAILURE
                             }
                         }
